feat: support disableSvgr
