+++ packages/builder/builder-shared/src/config.ts
@@ -81,6 +81,7 @@ export const getDefaultOutputConfig = (): NormalizedSharedOutputConfig => ({
   legalComments: 'linked',
   cleanDistPath: true,
   svgDefaultExport: 'url',
+  disableSvgr: false,
   disableCssExtract: false,
   disableMinimize: false,
   disableSourceMap: {
