+++ tests/e2e/builder/cases/svg/svg-assets/src/App.css
@@ -1,0 +1,5 @@
+#test-css {
+  width: 50px;
+  height: 50px;
+  background-image: url('./app.svg?url');
+}
