+++ tests/e2e/builder/cases/svg/svg-assets/src/App.jsx
@@ -1,0 +1,16 @@
+// eslint-disable-next-line no-unused-vars
+import React from 'react';
+import svgImg from './app.svg?url';
+import './App.css';
+
+function App() {
+  return (
+    <div>
+      <div id="test">Hello Builder!</div>
+      <img id="test-img" src={svgImg} />
+      <div id="test-css" />
+    </div>
+  );
+}
+
+export default App;
