+++ tests/e2e/builder/cases/svg/svg-assets/src/index.js
@@ -1,0 +1,7 @@
+// eslint-disable-next-line eslint-comments/disable-enable-pair
+/* eslint-disable no-undef */
+import React from 'react';
+import { render } from 'react-dom';
+import App from './App';
+
+render(React.createElement(App), document.getElementById('root'));
