+++ tests/e2e/builder/cases/svg/svg-default-export-component/src/App.jsx
