+++ tests/e2e/builder/cases/svg/svg-default-export-component/src/index.js
