+++ tests/e2e/builder/cases/svg/svg-external-react/src/App.jsx
