+++ tests/e2e/builder/cases/svg/svg-external-react/static/index.html
