+++ tests/e2e/builder/cases/svg/svg-url/src/index.js
