+++ tests/e2e/builder/cases/svg/svg.test.ts
@@ -1,49 +1,170 @@
 import { join } from 'path';
-import { readFileSync } from 'fs';
 import { expect } from '@modern-js/e2e/playwright';
-import { build } from '@scripts/shared';
+import { build, getHrefByEntryName } from '@scripts/shared';
 import { allProviderTest } from '@scripts/helper';
 
-allProviderTest('should preserve viewBox after svgo minification', async () => {
-  const fixture = join(__dirname, 'svgo-minify-view-box');
+const fixtures = __dirname;
+
+allProviderTest('svg (default)', async ({ page }) => {
   const buildOpts = {
-    cwd: fixture,
+    cwd: join(fixtures, 'svg'),
     entry: {
-      main: join(fixture, 'src/index.jsx'),
+      main: join(fixtures, 'svg', 'src/index.js'),
     },
   };
 
-  const builder = await build(buildOpts);
+  const builder = await build(buildOpts, {});
 
-  const files = await builder.unwrapOutputJSON();
-  const mainJs = Object.keys(files).find(
-    file => file.includes('/main.') && file.endsWith('.js'),
-  );
-  const content = readFileSync(mainJs!, 'utf-8');
+  await page.goto(getHrefByEntryName('main', builder.port));
 
-  expect(
-    content.includes('width:120,height:120,viewBox:"0 0 120 120"'),
-  ).toBeTruthy();
+  // test svgr（namedExport）
+  await expect(
+    page.evaluate(`document.getElementById('test-svg').tagName === 'svg'`),
+  ).resolves.toBeTruthy();
+
+  // test svg asset
+  await expect(
+    page.evaluate(
+      `document.getElementById('test-img').src.startsWith('data:image/svg')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  // test svg asset in css
+  await expect(
+    page.evaluate(
+      `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('url("data:image/svg')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  builder.close();
 });
 
-allProviderTest('should add id prefix after svgo minification', async () => {
-  const fixture = join(__dirname, 'svgo-minify-id-prefix');
+allProviderTest('svg (defaultExport component)', async ({ page }) => {
   const buildOpts = {
-    cwd: fixture,
+    cwd: join(fixtures, 'svg-default-export-component'),
     entry: {
-      main: join(fixture, 'src/index.jsx'),
+      main: join(fixtures, 'svg-default-export-component', 'src/index.js'),
     },
   };
 
-  const builder = await build(buildOpts);
+  const builder = await build(buildOpts, {
+    output: {
+      svgDefaultExport: 'component',
+    },
+  });
 
-  const files = await builder.unwrapOutputJSON();
-  const mainJs = Object.keys(files).find(
-    file => file.includes('/main.') && file.endsWith('.js'),
-  );
-  const content = readFileSync(mainJs!, 'utf-8');
+  await page.goto(getHrefByEntryName('main', builder.port));
 
-  expect(
-    content.includes('"linearGradient",{id:"idPrefix_svg__a"}'),
-  ).toBeTruthy();
+  await expect(
+    page.evaluate(`document.getElementById('test-svg').tagName === 'svg'`),
+  ).resolves.toBeTruthy();
+
+  builder.close();
+});
+
+allProviderTest('svg (url)', async ({ page }) => {
+  const buildOpts = {
+    cwd: join(fixtures, 'svg-url'),
+    entry: {
+      main: join(fixtures, 'svg-url', 'src/index.js'),
+    },
+  };
+
+  const builder = await build(buildOpts, {});
+
+  await page.goto(getHrefByEntryName('main', builder.port));
+
+  // test svg asset
+  await expect(
+    page.evaluate(
+      `document.getElementById('test-img').src.includes('static/svg/app')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  // test svg asset in css
+  await expect(
+    page.evaluate(
+      `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('static/svg/app')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  builder.close();
+});
+
+// It's an old bug when use svgr in css and external react.
+allProviderTest('svg (external react)', async ({ page }) => {
+  const buildOpts = {
+    cwd: join(fixtures, 'svg-external-react'),
+    entry: {
+      main: join(fixtures, 'svg-external-react', 'src/index.js'),
+    },
+  };
+
+  const builder = await build(buildOpts, {
+    output: {
+      externals: {
+        react: 'React',
+        'react-dom': 'ReactDOM',
+      },
+    },
+    html: {
+      template: './static/index.html',
+    },
+  });
+
+  await page.goto(getHrefByEntryName('main', builder.port));
+
+  // test svgr（namedExport）
+  await expect(
+    page.evaluate(`document.getElementById('test-svg').tagName === 'svg'`),
+  ).resolves.toBeTruthy();
+
+  // test svg asset
+  await expect(
+    page.evaluate(
+      `document.getElementById('test-img').src.startsWith('data:image/svg')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  // test svg asset in css
+  await expect(
+    page.evaluate(
+      `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('url("data:image/svg')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  builder.close();
+});
+
+allProviderTest('svg (disableSvgr)', async ({ page }) => {
+  const buildOpts = {
+    cwd: join(fixtures, 'svg-assets'),
+    entry: {
+      main: join(fixtures, 'svg-assets', 'src/index.js'),
+    },
+  };
+
+  const builder = await build(buildOpts, {
+    output: {
+      disableSvgr: true,
+    },
+  });
+
+  await page.goto(getHrefByEntryName('main', builder.port));
+
+  // test svg asset
+  await expect(
+    page.evaluate(
+      `document.getElementById('test-img').src.includes('static/svg/app')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  // test svg asset in css
+  await expect(
+    page.evaluate(
+      `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('static/svg/app')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  builder.close();
 });
