+++ tests/e2e/builder/cases/svg/svg/src/App.jsx
