+++ tests/e2e/builder/cases/svg/svg/src/index.js
